backend/requirements.txt
alembic
fastapi
passlib
uvicorn
sqlalchemy
pytest
pytest-xdist
pytest-asyncio
pytest-cov
requests
bcrypt
pyjwt
python-multipart
